Removed OnyxProject as we now have dedicated perms for checking project access. Replaced with functions for validating and retrieving fields directly
